# doc/openapi.yaml
openapi: 3.0.3
info:
  title: movie theater
  version: 1.0.11
servers:
  - url: https://petstore3.swagger.io/api/v3
...
paths:
...
  /movie/{movieId}:
    get:
      tags:
        - movie
      summary: Find movie by ID
      operationId: getMovieById
      parameters:
        - name: movieId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Movie'
        '400':
          description: Invalid ID supplied
        '404':
          description: Pet not found
    delete:
      tags:
        - movie
      summary: Deletes a movie
      operationId: deleteMovie
      parameters:
        - name: movieId
          in: path
          description: Movie id to delete
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '400':
          description: Invalid id
...
  /screening/{screeningId}:
    get:
      tags:
        - screening
      summary: Find screening by ID
      operationId: getScreeningById
      parameters:
        - name: screeningId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Screening'
        '400':
          description: Invalid ID supplied
        '404':
          description: Pet not found
    delete:
      tags:
        - screening
      summary: Deletes a screening
      operationId: deleteScreening
      parameters:
        - name: screeningId
          in: path
          description: screening id to delete
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '400':
          description: Invalid id
...
  /auditorium/{auditoriumId}:
    get:
      tags:
        - auditorium
      summary: Find auditorium by ID
      operationId: getauditoriumById
      parameters:
        - name: auditoriumId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Auditorium'
        '400':
          description: Invalid ID supplied
        '404':
          description: Pet not found
    delete:
      tags:
        - auditorium
      summary: Deletes a auditorium
      operationId: deleteAuditorium
      parameters:
        - name: auditoriumId
          in: path
          description: screening id to delete
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '400':
          description: Invalid id
...
  /seat/{seatId}:
    get:
      tags:
        - seat
      summary: Find auditorium by ID
      operationId: getSeatById
      parameters:
        - name: seatId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Auditorium'
        '400':
          description: Invalid ID supplied
        '404':
          description: Pet not found
    delete:
      tags:
        - seat
      summary: Deletes a seat
      operationId: deleteSeat
      parameters:
        - name: seatId
          in: path
          description: seat id to delete
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '400':
          description: Invalid id
...
  /employee/{employeeId}:
    get:
      tags:
        - employee
      summary: Find employee by ID
      operationId: getEmployeeById
      parameters:
        - name: employeeId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Employee'
        '400':
          description: Invalid ID supplied
        '404':
          description: Pet not found
    delete:
      tags:
        - employee
      summary: Deletes a employee
      operationId: deleteEmployee
      parameters:
        - name: employeeId
          in: path
          description: employee id to delete
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '400':
          description: Invalid id
...
components:
  schemas:
    Movie:
      required:
        - title
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        title:
          type: string
          example: Titanic
        director:
          type: string
          example: Robert
        cast:
          type: string
          example: Avana
        description:
          type: string
          example: alefij mzoej,zraoeif,,aeprjv,e
        duration_min:
          type: integer
          format: int64
          example: 120
    Screening:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        movie_title:
          type: string
          example: Titanic
        auditorium:
          type: string
          example: A
        duration_min:
          type: integer
          format: int64
          example: 120
    Employee:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        username:
          type: string
          example: loica
        password:
          type: string
          example: toky123
...
    Auditorium:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        auditorium_name:
          type: string
          example: A
...